Create paswd 2.0
Updated version of paswd.

Realized that , some sort of backup along with availability of the passwords online and too access with any device was necessary, hence i have modified the code and uploded it to replit , so now i can access my passwords from any device online. However , the stored passwords(encrypted) is still available on the cloud so it can be a potential vulnerability.